Redirect CPX NSPPE logs to kubectl logs
NSPPE logs for CPX will be redirected when tty is true for CPX container

Testing status:
Tested with this change and able to see NSPPE log in the CPX container

Note:
With this change, nsconfigd logs(Config commands) will also be logged. Assuming that logs access is restricted and so, it should be fine.